+++ tests/topology.cpp
@@ -1,32 +1,30 @@
 #include <catch.hpp>
 
-#define private public
-#include "chemfiles/Topology.hpp"
-#undef private
-
 #include "chemfiles.hpp"
 using namespace chemfiles;
 
 TEST_CASE("Use the Topology class", "[Topology]"){
 
     SECTION("Basic usage"){
-        auto topo = Topology(42);
-        CHECK(topo.atoms_.capacity() == 42);
+        auto topology = Topology();
 
-        topo.append(Atom("H"));
-        CHECK(topo[0].type() == Atom::ELEMENT);
-        CHECK(topo[0].name() == "H");
+        topology.append(Atom("H"));
+        CHECK(topology[0].type() == Atom::ELEMENT);
+        CHECK(topology[0].name() == "H");
 
-        topo.resize(100);
-        CHECK(topo.atoms_.size() == 100);
+        topology.append(Atom("H"));
+        CHECK(topology[0].type() == Atom::ELEMENT);
+        CHECK(topology[0].name() == "H");
+
+        topology.add_bond(0, 1);
+        CHECK(topology.bonds().size() == 1);
+        CHECK(topology.bonds()[0] == Bond(0, 1));
     }
 
     SECTION("Dummy topology provider"){
         auto topo = dummy_topology(42);
         CHECK(topo.natoms() == 42);
 
-        CHECK(topo.natom_types() == 1);
-        CHECK(topo.templates_[0] == Atom(Atom::UNDEFINED));
         CHECK(topo[10] == Atom(Atom::UNDEFINED));
     }
 
@@ -42,8 +40,6 @@ TEST_CASE("Use the Topology class", "[Topology]"){
         topo.add_bond(1, 4);
         topo.add_bond(2, 5);
         topo.add_bond(3, 5);
-
-        CHECK(topo.natom_types() == 2);
 
         CHECK(topo.bonds().size() == 4);
         CHECK(topo.isbond(0, 4));